+++ types/types.ts
@@ -1,0 +1,8 @@
+export type cardInfoType = {
+    url: string;
+    urlalt: string;
+    title: string;
+    text1: string;
+    text2: string;
+    text3: string;
+}
